틀린 문제에 대한 원인 파악

+++ CodingTest/src/main/java/codingtest/coupang/SCM_Back_end_Engineer_in_2022_2.java
@@ -16,6 +16,7 @@ public class SCM_Back_end_Engineer_in_2022_2 {
 	public static long[] solution(long n) {
         long[] answer = new long[2];
         List<Long> list = new ArrayList<Long>();
+        /**에라토스테네스 체로 값을 셋팅*/
         
         for(int i = 1 ; i * i <= n ; i++){
         	
@@ -23,6 +24,7 @@ public class SCM_Back_end_Engineer_in_2022_2 {
         		Long x = (long) i;
         		Long y = (long) (n / i);
         		
+        		/**셋팅된 값으로 소수 판별*/
         		if(x != y && isPrime(x, y)) {
         			list.add(x);
         			list.add(y);
@@ -48,19 +50,19 @@ public class SCM_Back_end_Engineer_in_2022_2 {
 		if(x == 1 || x == 0) return false;
 		if(y == 1 || y == 0) return false;
 		
-		for (int i = 2 ; i <= (int)Math.sqrt(x) ; i++) {
+		for (long i = 2 ; i <= Math.sqrt(x) ; i++) {
             if (x % i == 0) {
                return false;
             }//end if()
         }//end for()
 
-		for (int i = 2 ; i <= (int)Math.sqrt(y) ; i++) {
+		for (long i = 2 ; i <= Math.sqrt(y) ; i++) {
             if (y % i == 0) {
                return false;
             }//end if()
         }//end for()
 		
-         return true;
+        return true;
 	}//end isPrime()
 
 }//end class()
